Complete JSDoc for Montage core constructor
Consolidates the paths and bindings code into core so that JSDoc can
pick up their annotations. Also includes documentation for methods
imported from Collections for property change notification.
