+++ test/base/abstract-number-field-spec.js
@@ -1,7 +1,6 @@
 /*global describe, it, expect, beforeEach, afterEach */
 var Montage = require("montage").Montage,
     AbstractNumberField = require("montage/ui/base/abstract-number-field").AbstractNumberField,
-    Bindings = require("montage/core/bindings").Bindings,
     MockDOM = require("mocks/dom"),
     MockComponent = require("mocks/component"),
     MockEvent = require("mocks/event");
